pyproject.toml
[project]
name = "recomsys-etfs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gradio>=5.34.2",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.21",
    "langchain-community>=0.3.24",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "langchain-chroma>=0.2.4",
    "kiwipiepy>=0.21.0",
    "langfuse>=3.0.8",
    "langchain-google-genai>=2.1.5",
    "rank-bm25>=0.2.2",
]
